+++ test/Classes/ReviewsTest.php
@@ -2,7 +2,7 @@
 
 namespace Bazaarvoice;
 
-require_once 'bvseosdk.php';
+require_once 'src/bvseosdk.php';
 require_once 'test/config.php';
 
 /**
